Fixing bigint to int for test

+++ cpg-language-rust/src/main/kotlin/de/fraunhofer/aisec/cpg/frontends/rust/ExpressionHandler.kt
@@ -256,7 +256,11 @@ class ExpressionHandler(frontend: RustLanguageFrontend) :
                 else -> null
             }
 
-        return newLiteral(value, typ?.let { language.builtInTypes[it] } ?: unknownType(), raw)
+        return newLiteral(
+            value.toInt(),
+            typ?.let { language.builtInTypes[it] } ?: unknownType(),
+            raw,
+        )
     }
 
     fun handleCallExpr(callExpr: RsCallExpr): Call {
